flexible hpcg gpu check
